gServices/Product/requirements.txt
aiopg
grpcio
grpcio-tools
piccolo[postgres]
protobuf==3.20.*

opentelemetry-api==1.15.0
opentelemetry-exporter-jaeger==1.15.0
opentelemetry-exporter-jaeger-proto-grpc==1.15.0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-instrumentation-grpc==0.36b0
opentelemetry-sdk==1.15.0

grpcio-health-checking
grpcio-reflection

locust
grpc-interceptor
